Refactor testscript framework
